Updated hourly allocation data

+++ client/src/erp-modules/payroll/workview/paygroup-detail-table/workview-tabs/hourly/HourlyAllocation.jsx
@@ -70,7 +70,7 @@ const HourlyAllocation = ({ company, closestRecord, groupId, payrunOption }) => 
 				updatedRec[totalColumnKey] = 0;
 				colKeys.forEach((key) => {
 					if (key && key !== totalColumnKey) {
-						updatedRec[totalColumnKey] += parseInt(updatedRec[key]);
+						updatedRec[totalColumnKey] += parseFloat(updatedRec[key]);
 					}
 				});
 				updatedRec.companyName = company;
